data-processing/requirements.txt
# OncoHotspot ETL Pipeline Requirements

# Core dependencies
requests>=2.28.0
numpy>=1.21.0
pyyaml>=6.0

# Optional for advanced features
pandas>=1.3.0  # For data manipulation
schedule>=1.1.0  # For task scheduling
redis>=4.3.0  # For caching (optional)

# Development dependencies
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0